task(java): format and update code
