+++ packages/java/neo-auth-service/src/main/java/org/neocoder/services/auth/AuthService.java
@@ -1,4 +1,4 @@
-package org.neocoder.authservice;
+package org.neocoder.services.auth;
 
 import org.mindrot.jbcrypt.BCrypt;
