+++ packages/java/neo-auth-service/src/main/java/org/neocoder/services/auth/TokenService.java
@@ -1,0 +1,94 @@
+package org.neocoder.services.auth;
+
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.JWTVerifier;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.interfaces.DecodedJWT;
+
+import java.time.Instant;
+import java.util.Date;
+import java.util.List;
+
+import org.neocoder.services.auth.exception.InvalidTokenException;
+
+public class TokenService {
+    private String secret;
+    private String issuer;
+    private int expTime;
+    private static final String USERNAME = "";
+
+    public TokenService(String secret, String issuer, int expTime) {
+        this.secret = secret;
+        this.issuer = issuer;
+        this.expTime = expTime;
+    }
+
+    private String generate(String user, List<String> roles) {
+        Instant now = Instant.now();
+        Algorithm algorithm = Algorithm.HMAC256(secret);
+
+        return JWT.create()
+                .withIssuer(issuer)
+                .withIssuedAt(Date.from(now))
+                .withExpiresAt(Date.from(now.plusSeconds(expTime)))
+                .withClaim(USERNAME, user)
+                .withClaim("roles", roles)
+                .sign(algorithm);
+    }
+
+    public Claims decode(String token) throws InvalidTokenException {
+        try {
+            Algorithm algorithm = Algorithm.HMAC256(secret);
+            JWTVerifier verifier = JWT.require(algorithm)
+                    .build();
+
+            DecodedJWT decodedJWT = verifier.verify(token);
+
+            long exp = decodedJWT.getExpiresAt().getTime() / 1000; // Convert milliseconds to seconds
+            long iat = decodedJWT.getIssuedAt().getTime() / 1000;
+            String iss = decodedJWT.getIssuer();
+            String username = decodedJWT.getClaim(USERNAME).asString();
+            List<String> roles = decodedJWT.getClaim("roles").asList(String.class);
+
+            return new Claims(exp, iat, iss, username, roles);
+
+        } catch (Exception e) {
+            throw new InvalidTokenException("Invalid JWT token", e);
+        }
+    }
+
+    private Claims verify(String token) throws InvalidTokenException {
+        try {
+            Algorithm algorithm = Algorithm.HMAC256(secret);
+            JWTVerifier verifier = JWT.require(algorithm)
+                    .build();
+
+            DecodedJWT decodedJWT = verifier.verify(token);
+
+            long exp = decodedJWT.getExpiresAt().getTime() / 1000; // Convert milliseconds to seconds
+            long iat = decodedJWT.getIssuedAt().getTime() / 1000;
+            String iss = decodedJWT.getIssuer();
+            String username = decodedJWT.getClaim(USERNAME).asString();
+            List<String> roles = decodedJWT.getClaim("roles").asList(String.class);
+
+            Claims claims = new Claims(exp, iat, iss, username, roles);
+
+            if (claims.isExpired()) {
+                throw new InvalidTokenException("Token has expired");
+            }
+
+            if (!claims.isIssuedBy(iss)) {
+                throw new InvalidTokenException("Invalid issuer");
+            }
+
+            return claims;
+        } catch (Exception e) {
+            throw new InvalidTokenException("Invalid token", e);
+        }
+    }
+
+    public String refresh(String token) {
+        Claims payload = verify(token);
+        return generate(payload.username, payload.roles);
+    }
+}
