+++ packages/java/neo-auth-service/src/main/java/org/neocoder/services/auth/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package org.neocoder.authservice.exception;
+package org.neocoder.services.auth.exception;
 
 public class InvalidTokenException extends RuntimeException{
     public InvalidTokenException(String message){
